+++ CHANGELOG.rst
@@ -24,6 +24,7 @@ Added
     - ``get_project_by_uuid()`` and ``get_user_by_username()`` template tags
     - ``get_remote_icon()`` template tag (#3)
     - Predicates in rules for handling remote projects (#3)
+    - ``ProjectModifyPermissionMixin`` for access control for remote projects (#3)
 - **Timeline**
     - ``RemoteSite`` model in ``api.get_event_description()`` (#3)